% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_models.R
\name{candidate_models}
\alias{candidate_models}
\title{Candidate Models for regression}
\usage{
candidate_models(
  variable_names = paste0("x", 1:10),
  model_size = 3,
  groups = NA,
  max_mix = 1,
  dv_name = "y",
  output_short = FALSE
)
}
\arguments{
\item{variable_names-}{series of varaible names [paste0("x", 1:10)]}

\item{model_size-}{number of independent variables in the model formula. Default [3].}

\item{groups-}{same size as model_size variable.
Defines the potential grouping of the variables. Default  [NA].}

\item{max_mix-}{the maximum number of repeated variables from the same group in the model. Default  [1]}

\item{dv_name-}{name of the dependent variable used in the formula.
Default  ["y"].}

\item{output_short-}{if the output should be limited to the formulas only.   [FALSE]}
}
\description{
Function creates a candate model formulas
on the basis of the varaible names.
}
\examples{

cand_mod0 <- candidate_models(variable_names = paste0("x", 1:100),
groups = c(rep(1,30),rep(2,30), rep(3,20), rep(4,20)),
model_size = 4,
max_mix = c(1))

dim(cand_mod0$candidate_all)
dim(cand_mod0$candidate_by_grouping)

cand_mod1 <- candidate_models(variable_names = paste0("x", 1:100),
                              groups = c(rep(1,30),rep(2,30), rep(3,20), rep(4,20)),
                              model_size = 3,
                              max_mix = c(1),
                              output_short = FALSE)

dim(cand_mod1$candidate_all)
dim(cand_mod1$candidate_by_grouping)

cand_mod2 <- candidate_models(variable_names = paste0("x", 1:100),
                              groups = c(rep(1,30),rep(2,30), rep(3,40)),
                              model_size = 2,
                              max_mix = c(1),
                              output_short = FALSE)


dim(cand_mod2$candidate_all)
dim(cand_mod2$candidate_by_grouping)




}
